requirements.txt
streamlit
pandas
requests
tqdm
tensorflow
faiss-cpu
scikit-learn
numpy
matplotlib
pillow